♻️ refactor : 우승자 확인 및 출력하는 메서드 모듈화
우승자를 확인하고 출력해주는 resultGame 모듈화 완성.

ex) Fixes : #45 Related to : #34, #23
